feat: add exposure compensation
The exposure compensation values are now represented as `Float` instead of `Int`

+++ camposer/src/iosMain/kotlin/com/ujizin/camposer/CameraPreview.ios.kt
@@ -34,7 +34,7 @@ internal actual fun CameraPreviewImpl(
     flashMode: FlashMode,
     scaleType: ScaleType,
     enableTorch: Boolean,
-    exposureCompensation: Int, // TODO
+    exposureCompensation: Float,
     zoomRatio: Float,
     imageAnalyzer: ImageAnalyzer?, // TODO
     implementationMode: ImplementationMode,
